ZXBSInstaller v0.0.1-beta6: Show and install old versions ZXBasicStudio v1.7.0
