+++ ZXBSInstaller.Log/ZXBSInstaller.Log.csproj
@@ -11,19 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="InstallerResources\zxbasic.scr.0.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
     <Content Include="InstallerResources\zxbasic.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="InstallerResources\zxbasic.scr.3.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="InstallerResources\zxbasic.scr.2.png">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="InstallerResources\zxbasic.scr.1.png">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="InstallerResources\zxbsinstaller.png">
