--- ZXBSInstaller/Controls/ConfigControl.axaml
@@ -1,24 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 Width="640" Height="250"
-			 x:Class="ZXBSInstaller.Controls.ConfigControl">
-	<Border BorderBrush="White" BorderThickness="1" CornerRadius="10" Background="Black">
-		<Grid Grid.RowDefinitions="*,4,Auto" Margin="16">
-			<StackPanel Orientation="Vertical">
-				<TextBlock FontSize="16" FontWeight="Bold" Text="Configuration options"/>
-				<TextBlock Margin="0,16,0,0" Text="Base path for external tools"/>
-				<Grid Grid.ColumnDefinitions="*,Auto" Margin="0,4,0,0">
-					<TextBox x:Name="txtToolsBasePath"/>
-					<Button x:Name="btnBrowseToolsBasePath" Grid.Column="1" Content="..." Margin="4,0,0,0" Padding="2" Click="btnBrowseToolsBasePath_Click"/>
-				</Grid>
-				<CheckBox x:Name="chkOnlyStableVersions" Margin="0,16,0,0" Content="Show only stable versions (no beta)"/>
-			</StackPanel>
-			<StackPanel Orientation="Horizontal" Grid.Row="2" HorizontalAlignment="Right">
-				<Button x:Name="btnCancel" Content="Cancel" Width="120" Margin="4" Click="btnCancel_Click"/>
-				<Button x:Name="btnSave" Content="Save" Width="120" Margin="4" Click="btnSave_Click"/>
-			</StackPanel>
-		</Grid>
-	</Border>
-</UserControl>
