--- ZXBSInstaller/Controls/ScreenShotControl.axaml
@@ -1,10 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-			 x:Class="ZXBSInstaller.Controls.ScreenShotControl">
-	<Border BorderBrush="White" BorderThickness="1" Margin="4" Width="202" Height="202">
-		<Image x:Name="imgScreenShot" Stretch="Uniform" Width="200" Height="200" Margin="1"/>
-	</Border>
-</UserControl>
