--- ZXBSInstaller/Controls/ToolsListItemControl.axaml
@@ -1,14 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-			 x:Class="ZXBSInstaller.ToolsListItemControl">
-	<Grid Grid.ColumnDefinitions="32,4,*">
-		<Image x:Name="imgIcon" Width="32" Stretch="Uniform" VerticalAlignment="Top" />
-		<StackPanel Orientation="Vertical" Grid.Column="2" Margin="0">
-			<TextBlock x:Name="txtName" FontWeight="Bold" TextWrapping="Wrap" Margin="0"/>
-			<TextBlock x:Name="txtVersion" Foreground="LightGray" Margin="0"/>
-		</StackPanel>
-	</Grid>
-</UserControl>
