--- ZXBSInstaller/Controls2/ToolItemControl.axaml
@@ -1,24 +1,0 @@
-<UserControl xmlns="https://github.com/avaloniaui"
-			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-			 xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-			 xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-			 Width="360" Height="150"
-			 x:Class="ZXBSInstaller.Controls2.ToolItemControl"
-			 FontSize="12">
-	<Grid Margin="8"  Background="#242424">
-		<Grid Grid.ColumnDefinitions="Auto,4,*,4,Auto" Grid.RowDefinitions="Auto,*,Auto,Auto" Margin="4">
-			<Image x:Name="imgIcon" Grid.RowSpan="3" Width="64" Stretch="Uniform" VerticalAlignment="Top" />
-			<TextBlock x:Name="txtName" Grid.Column="2" VerticalAlignment="Top" TextWrapping="Wrap" FontSize="14" FontWeight="Bold">Name</TextBlock>
-			<TextBlock x:Name="txtDescription" Grid.Column="2" Grid.Row="1" VerticalAlignment="Top" TextWrapping="Wrap">Description</TextBlock>
-			<CheckBox x:Name="chkSelect" Grid.Column="4" VerticalAlignment="Top" IsCheckedChanged="chkSelect_IsCheckedChanged"/>
-
-			<StackPanel Orientation="Vertical" Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="2" VerticalAlignment="Top" Margin="0,0,50,0">
-				<TextBlock x:Name="txtPath" TextWrapping="Wrap" IsVisible="False">Path</TextBlock>
-				<TextBlock x:Name="txtActual">Installed version:</TextBlock>
-				<TextBlock x:Name="txtLatest">Latest version:</TextBlock>
-			</StackPanel>
-
-			<Button x:Name="btnAllVersions" Grid.ColumnSpan="5" Grid.Row="2" HorizontalAlignment="Right" VerticalAlignment="Bottom" IsVisible="False">...</Button>
-		</Grid>
-	</Grid>
-</UserControl>
