+++ ZXBSInstaller/ZXBSInstaller.csproj
@@ -31,16 +31,4 @@
     <ProjectReference Include="..\MsBox.Avalonia\MsBox.Avalonia.csproj" />
     <ProjectReference Include="..\ZXBSInstaller.Log\ZXBSInstaller.Log.csproj" />
   </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Controls\ScreenShotControl.axaml.cs">
-      <DependentUpon>ScreenShotControl.axaml</DependentUpon>
-    </Compile>
-    <Compile Update="Controls\ToolsListItemControl.axaml.cs">
-      <DependentUpon>ToolsListItemControl.axaml</DependentUpon>
-    </Compile>
-    <Compile Update="Controls\VersionControl.axaml.cs">
-      <DependentUpon>VersionControl.axaml</DependentUpon>
-    </Compile>
-  </ItemGroup>
 </Project>
