+++ ZXBStudio/ZXBasicStudio.csproj
@@ -13,7 +13,7 @@
     <Title>ZX Basic Studio</Title>
     <AssemblyVersion></AssemblyVersion>
     <Deterministic>False</Deterministic>
-    <Version>1.6.0.5</Version>
+    <Version>1.7.0.0</Version>
   </PropertyGroup>
   <ItemGroup>
     <AvaloniaXaml Remove="LanguageDefinitions\**" />
