+++ ZXBasicStudio.sln
@@ -17,9 +17,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Rom128Reconstructor", "Rom1
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MsBox.Avalonia", "MsBox.Avalonia\MsBox.Avalonia.csproj", "{1AC800DC-80F6-45C2-AF8A-861C31FB544D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZXBasicStudioSite", "ZXBasicStudioSite\ZXBasicStudioSite.csproj", "{2493E207-2E56-48AB-8064-0E7A805F7A62}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Elementos de la solución", "Elementos de la solución", "{C949A587-5005-4308-BCFD-85831B080589}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Docs", "Docs", "{C949A587-5005-4308-BCFD-85831B080589}"
 	ProjectSection(SolutionItems) = preProject
 		LICENSE.txt = LICENSE.txt
 		README.md = README.md
@@ -37,6 +35,12 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZXBSInstaller", "ZXBSInstaller\ZXBSInstaller.csproj", "{67768F7F-169A-420F-9C50-335FB9885EC1}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ZXBSInstaller.Log", "ZXBSInstaller.Log\ZXBSInstaller.Log.csproj", "{A31A06FF-7FAD-4D57-A281-8034BA761669}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "IDE", "IDE", "{E148229E-C8ED-4DD7-9768-9641E279DD32}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Installer", "Installer", "{1613C3B0-B7D2-4FE5-A96A-A568C370DE84}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Emulator", "Emulator", "{6AB4C5D0-AB60-49F0-A349-E46971736749}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -102,14 +106,6 @@ Global
 		{1AC800DC-80F6-45C2-AF8A-861C31FB544D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{1AC800DC-80F6-45C2-AF8A-861C31FB544D}.Release|x64.ActiveCfg = Release|Any CPU
 		{1AC800DC-80F6-45C2-AF8A-861C31FB544D}.Release|x64.Build.0 = Release|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Debug|x64.Build.0 = Debug|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Release|x64.ActiveCfg = Release|Any CPU
-		{2493E207-2E56-48AB-8064-0E7A805F7A62}.Release|x64.Build.0 = Release|Any CPU
 		{D99C36AE-072A-4885-9585-CFD78FC6763E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D99C36AE-072A-4885-9585-CFD78FC6763E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D99C36AE-072A-4885-9585-CFD78FC6763E}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -145,6 +141,18 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{F65553BC-B86F-4125-B5BA-EBBCC223100B} = {E148229E-C8ED-4DD7-9768-9641E279DD32}
+		{726F8E04-840F-4F15-935D-429D81AC0311} = {E148229E-C8ED-4DD7-9768-9641E279DD32}
+		{F91A239A-F9AB-42A3-AECD-9AB844A5B908} = {6AB4C5D0-AB60-49F0-A349-E46971736749}
+		{98F26A78-7E51-49D1-8B09-7A85F2B28D29} = {E148229E-C8ED-4DD7-9768-9641E279DD32}
+		{3EFA2E44-4D1E-4212-B6BB-9F34CD228772} = {6AB4C5D0-AB60-49F0-A349-E46971736749}
+		{8B15A1B8-5C8A-4F81-A50B-834687EEE040} = {6AB4C5D0-AB60-49F0-A349-E46971736749}
+		{1AC800DC-80F6-45C2-AF8A-861C31FB544D} = {E148229E-C8ED-4DD7-9768-9641E279DD32}
+		{D99C36AE-072A-4885-9585-CFD78FC6763E} = {6AB4C5D0-AB60-49F0-A349-E46971736749}
+		{67768F7F-169A-420F-9C50-335FB9885EC1} = {1613C3B0-B7D2-4FE5-A96A-A568C370DE84}
+		{A31A06FF-7FAD-4D57-A281-8034BA761669} = {1613C3B0-B7D2-4FE5-A96A-A568C370DE84}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {58A8B264-A797-4839-A806-C8FEEFD4F80E}
